requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0

# Data fetching
requests>=2.28.0
yfinance>=0.2.18

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Statistical analysis
scipy>=1.9.0

# Development and testing (optional)
pytest>=7.0.0
pytest-mock>=3.10.0
black>=22.0.0
flake8>=5.0.0